+++ tests/Extractor/MessageExtractorTest.php
@@ -13,11 +13,13 @@ use FormatPHP\Extractor\Parser\DescriptorParserInterface;
 use FormatPHP\Extractor\Parser\ParserErrorCollection;
 use FormatPHP\Test\TestCase;
 use FormatPHP\Util\FileSystemHelper;
+use FormatPHP\Util\FormatHelper;
 use FormatPHP\Util\Globber;
 use Generator;
-use Hamcrest\Type\IsResource;
+use Hamcrest\Core\IsInstanceOf;
 use Psr\Log\LoggerInterface;
 use Psr\Log\NullLogger;
+use stdClass;
 
 use function json_decode;
 use function ob_end_clean;
@@ -44,7 +46,7 @@ class MessageExtractorTest extends TestCase
         $logger = $this->mockery(LoggerInterface::class);
         $logger->expects()->warning('Could not find files', ['files' => ['foo', 'bar', 'baz']]);
 
-        $extractor = new MessageExtractor($options, $logger, $globber, $file);
+        $extractor = new MessageExtractor($options, $logger, $globber, $file, new FormatHelper($file));
         $extractor->process(['foo', 'bar', 'baz']);
     }
 
@@ -59,6 +61,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -114,6 +117,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -160,6 +164,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -192,6 +197,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -248,6 +254,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -288,13 +295,14 @@ class MessageExtractorTest extends TestCase
     {
         $logger = new NullLogger();
         $options = new MessageExtractorOptions();
-        $options->format = __DIR__ . '/format.php';
+        $options->format = __DIR__ . '/fixtures/formatter.php';
 
         $extractor = new MessageExtractor(
             $options,
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -346,6 +354,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -364,6 +373,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -383,9 +393,16 @@ class MessageExtractorTest extends TestCase
 
         $file = $this->mockery(FileSystemHelper::class);
         $file->shouldReceive('getContents')->andReturn('nothing of consequence');
-        $file->expects()->writeContents('en-US.json', "{}\n");
+        $file->expects()->writeJsonContents('en-US.json', new IsInstanceOf(stdClass::class));
 
-        $extractor = new MessageExtractor($options, $logger, new Globber(new FileSystemHelper()), $file);
+        $extractor = new MessageExtractor(
+            $options,
+            $logger,
+            new Globber(new FileSystemHelper()),
+            $file,
+            new FormatHelper(new FileSystemHelper()),
+        );
+
         $extractor->process([__DIR__ . '/Parser/Descriptor/fixtures/*.php']);
     }
 
@@ -399,7 +416,7 @@ class MessageExtractorTest extends TestCase
 
         $file = $this->mockery(FileSystemHelper::class);
         $file->expects()->getContents($path)->andThrows($exception);
-        $file->expects()->writeContents(new IsResource(), "{}\n");
+        $file->expects()->writeJsonContents('php://output', new IsInstanceOf(stdClass::class));
 
         $logger = $this->mockery(LoggerInterface::class);
         $logger->shouldReceive('debug')->with(
@@ -411,7 +428,14 @@ class MessageExtractorTest extends TestCase
             ['exception' => $exception],
         );
 
-        $extractor = new MessageExtractor($options, $logger, new Globber(new FileSystemHelper()), $file);
+        $extractor = new MessageExtractor(
+            $options,
+            $logger,
+            new Globber(new FileSystemHelper()),
+            $file,
+            new FormatHelper(new FileSystemHelper()),
+        );
+
         $extractor->process([$path]);
     }
 
@@ -433,7 +457,13 @@ class MessageExtractorTest extends TestCase
             ['file' => $path],
         );
 
-        $extractor = new MessageExtractor($options, $logger, new Globber(new FileSystemHelper()), $file);
+        $extractor = new MessageExtractor(
+            $options,
+            $logger,
+            new Globber(new FileSystemHelper()),
+            $file,
+            new FormatHelper(new FileSystemHelper()),
+        );
 
         $this->expectException(UnableToProcessFileException::class);
         $this->expectExceptionMessage('something bad happened');
@@ -452,6 +482,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -543,6 +574,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         ob_start();
@@ -634,6 +666,7 @@ class MessageExtractorTest extends TestCase
             $logger,
             new Globber(new FileSystemHelper()),
             new FileSystemHelper(),
+            new FormatHelper(new FileSystemHelper()),
         );
 
         $this->expectException(InvalidArgumentException::class);