+++ tests/Util/fixtures/reader-closure-invalid-04.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+use FormatPHP\ConfigInterface;
+use FormatPHP\Intl\LocaleInterface;
+use FormatPHP\MessageCollection;
+
+return function (ConfigInterface $config, $data, LocaleInterface $localeResolved): MessageCollection
+{
+    return new MessageCollection();
+};