+++ tests/Util/fixtures/writer-closure-invalid-03.php
@@ -1,0 +1,10 @@
+<?php
+
+declare(strict_types=1);
+
+use FormatPHP\Extractor\MessageExtractorOptions;
+
+return function ($collection, MessageExtractorOptions $options): array
+{
+    return [];
+};